PaytollControl/requirements.txt
google-genai>=1.0.0
google-cloud-vision>=3.5.0
pypdfium2>=4.0.0
Pillow>=10.0.0
xlwings>=0.30.0
pytest>=8.0.0
openpyxl>=3.0.0
numpy>=1.20.0

